修复回帖数量和用户查看公告bug

+++ technical_forum_backend/technical_forum_service/src/main/java/jee/sanda/forum/entity/ForumPost.java
@@ -63,5 +63,5 @@ public class ForumPost {
 //    @JoinColumn(name="forum_post_id")
 //    private Set<ForumPostDetail> forumPostDetails;
     @Transient
-    private Long commentQuantity;
+    private Integer commentQuantity;
 }
