+++ technical_forum_backend/technical_forum_service/src/main/java/jee/sanda/forum/repository/ForumPostDetailRepository.java
@@ -19,5 +19,5 @@ public interface ForumPostDetailRepository extends JpaRepository<ForumPostDetail
     //计算回帖数量
     @Query(value="SELECT count(*) FROM forum_post_detail WHERE forum_post_id=?1",nativeQuery = true)
     @Modifying
-    Long countByPostId(Long PostId);
+    Integer countByPostId(Long PostId);
 }
