+++ technical_forum_backend/technical_forum_service/src/main/java/jee/sanda/forum/service/ForumPostService.java
@@ -79,5 +79,5 @@ public interface ForumPostService {
      * @param forumPostId
      * @return
      */
-    Long countCommentQuantity(Long forumPostId);
+    Integer countCommentQuantity(Long forumPostId);
 }
